Developer/fix logic

+++ ShoppingAPI/Controllers/ShoppingDeliveryAddressController.cs
@@ -1,7 +1,12 @@
 ﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ShoppingAPI.Common;
+using ShoppingAPI.Common.Config;
+using ShoppingAPI.Common.Extensions;
 using ShoppingAPI.Common.Models;
 using ShoppingAPI.Data.Models;
+using ShoppingAPI.Model;
 using ShoppingAPI.Services.Interfaces;
 using System.IdentityModel.Tokens.Jwt;
 using System.Net;
@@ -11,57 +16,101 @@ namespace ShoppingAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    [Authorize(Roles = "Admin,SuperAdmin")]
+    [Authorize]
     public class ShoppingDeliveryAddressController : ControllerBase
     {
         private readonly IShoppingDeliveryAddressServices shoppingDeliveryAddressServices;
-        public ShoppingDeliveryAddressController(IShoppingDeliveryAddressServices shoppingDeliveryAddressServices)
+        private int UserId = -1;
+        private string roleName = "Guest";
+        public ShoppingDeliveryAddressController(IShoppingDeliveryAddressServices shoppingDeliveryAddressServices, IHttpContextAccessor httpContextAccessor)
         {
             this.shoppingDeliveryAddressServices = shoppingDeliveryAddressServices;
+            this.UserId = int.Parse(httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value);
+            this.roleName = httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.Role).Value;
         }
-        [HttpGet]
-        public async Task<IActionResult> ShoppingDeliveryAddresses()
+
+        [HttpGet("[Action]")]
+        public async Task<IActionResult> getShoppingDeliveryAddresses(int? UserId, int? page, int? pageSize)
         {
-            var categories = await shoppingDeliveryAddressServices.GetShoppingDeliveryAddressesAsync();
-            return Ok(new ResultApi
+            if (UserId == null)
+                UserId = this.UserId;
+            else if (!Library.isAdmin(roleName) && UserId != this.UserId)
+                return Unauthorized();
+
+            if (page == null)
+                page = PagingSettingsConfig.pageDefault;
+            if (pageSize == null)
+                pageSize = PagingSettingsConfig.pageSize;
+
+            var DeliveryAddress = await shoppingDeliveryAddressServices
+                .GetShoppingDeliveryAddresses()
+                .Where(x => x.UserId == this.UserId)
+                .OrderByDescending(x => x.Id)
+                .ToPagedList(page.Value, pageSize.Value);
+
+            return Ok(new ResponseWithPaging
             {
                 Status = (int)HttpStatusCode.OK,
                 Success = true,
-                Data = categories
+                Data = DeliveryAddress,
+                PageCount = DeliveryAddress.PageCount,
+                PageNumber = DeliveryAddress.PageNumber,
+                TotalItems = DeliveryAddress.TotalItemCount
             });
         }
-        [HttpGet("{id}"), AllowAnonymous]
-        public async Task<IActionResult> ShoppingDeliveryAddress(int id)
+
+        //Get ShoppingDeliveryAddress from Id
+        [HttpGet("[Action]/{id}")]
+        public async Task<IActionResult> getShoppingDeliveryAddress(int id)
         {
-            string roles = User.Claims.FirstOrDefault(x => x.Type.Equals(ClaimTypes.Role))?.Value ?? "";
-            var UserId = User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier).Value;
             var shoppingDeliveryAddress = await shoppingDeliveryAddressServices.GetShoppingDeliveryAddressAsync(id);
+
+            //Check ShoppingDeliveryAddress Exists
             if (shoppingDeliveryAddress != null)
             {
-                if (roles.Equals("SuperAdmin") || roles.Equals("Admin") || UserId.Equals(shoppingDeliveryAddress.UserId.ToString()))
-                    return Ok(new ResultApi
+                //check if admin or user for Address user
+                if (Library.isAdmin(roleName) || this.UserId == shoppingDeliveryAddress.UserId)
+                    return Ok(new ResponseApi
                     {
                         Status = (int)HttpStatusCode.OK,
                         Data = shoppingDeliveryAddress,
                         Success = true
                     });
-                return NotFound(new ResultApi
-                {
-                    Status = (int)HttpStatusCode.NotFound,
-                    Success = false,
-                    Message = new[] { "Not found shoppingDeliveryAddress User" }
-                });
+                return Unauthorized();
+
             }
-            return Unauthorized();
+            return NotFound(new ResponseApi
+            {
+                Status = (int)HttpStatusCode.NotFound,
+                Success = false,
+                Message = new[] { "Not found shoppingDeliveryAddress User" }
+            });
         }
 
-        [HttpPost]
-        public async Task<IActionResult> shoppingDeliveryAddress(ShoppingDeliveryAddress shoppingDelivery)
+        //Insert ShoppingDelivery Address by User
+        [HttpPost("[Action]")]
+        public async Task<IActionResult> insertShoppingDeliveryAddress(ShoppingDeliveryAddressModel shoppingDeliveryAddressModel)
         {
             if (ModelState.IsValid)
             {
+                var shoppingDelivery = new ShoppingDeliveryAddress
+                {
+                    IsTrash = shoppingDeliveryAddressModel.IsTrash,
+                    Address = shoppingDeliveryAddressModel.Address,
+                    FullName = shoppingDeliveryAddressModel.FullName,
+                    IsDefault = shoppingDeliveryAddressModel.IsDefault,
+                    PhoneNumber = shoppingDeliveryAddressModel.PhoneNumber,
+                    UserId = shoppingDeliveryAddressModel.UserId,
+                };
+
+                if (Library.isAdmin(roleName))
+                {
+                    shoppingDelivery.UserId = this.UserId;
+                }
+
                 await shoppingDeliveryAddressServices.InsertShoppingDeliveryAddress(shoppingDelivery);
-                return Ok(new ResultApi
+
+                return Ok(new ResponseApi
                 {
                     Status = (int)HttpStatusCode.OK,
                     Success = true,
@@ -72,39 +121,53 @@ namespace ShoppingAPI.Controllers
             }
             return BadRequest();
         }
+        //Update Shopping DeliveryAddress by User
 
-        [HttpPut("InfomationUser")]
-        public async Task<IActionResult> PutInfomationUser(ShoppingDeliveryAddress shoppingDelivery)
+        [HttpPut("[Action]")]
+        public async Task<IActionResult> editInfomationUser(ShoppingDeliveryAddressModel shoppingDeliveryAddressModel)
         {
             if (ModelState.IsValid)
             {
-                var shoppingDeliveryAddressDb = await shoppingDeliveryAddressServices.GetShoppingDeliveryAddressAsync(shoppingDelivery.Id);
-
-                shoppingDeliveryAddressDb.PhoneNumber = shoppingDelivery.PhoneNumber;
-                shoppingDeliveryAddressDb.Address = shoppingDelivery.Address;
-
-                await shoppingDeliveryAddressServices.UpdateShoppingDeliveryAddress(shoppingDeliveryAddressDb);
-
-                return Ok(new ResultApi
+                if (this.UserId == shoppingDeliveryAddressModel.UserId || Library.isAdmin(roleName))
                 {
-                    Status = (int)HttpStatusCode.OK,
-                    Success = true,
-                    Message = new[] { "Edit success" },
-                    Data = shoppingDeliveryAddressDb
-                });
+                    var shoppingDeliveryAddressDb = await shoppingDeliveryAddressServices
+                        .GetShoppingDeliveryAddressAsync(shoppingDeliveryAddressModel.Id);
+
+                    shoppingDeliveryAddressDb.PhoneNumber = shoppingDeliveryAddressModel.PhoneNumber;
+                    shoppingDeliveryAddressDb.Address = shoppingDeliveryAddressModel.Address;
+
+                    await shoppingDeliveryAddressServices
+                        .UpdateShoppingDeliveryAddress(shoppingDeliveryAddressDb);
+
+                    return Ok(new ResponseApi
+                    {
+                        Status = (int)HttpStatusCode.OK,
+                        Success = true,
+                        Message = new[] { "Edit success" },
+                        Data = shoppingDeliveryAddressDb
+                    });
+                }
             }
-            return BadRequest();
+            return Unauthorized();
         }
 
-        [HttpDelete("InfomationUser")]
-        public async Task<IActionResult> DeleteShoppingDeliveryAddress(int id)
+        //Delete deleteShoppingDeliveryAddress
+
+        [HttpDelete("[Action]")]
+        public async Task<IActionResult> deleteShoppingDeliveryAddress(int id)
         {
-            await shoppingDeliveryAddressServices.DeleteShoppingDeliveryAddress(id);
-            return Ok(new ResultApi
+            var shoppingbyUser = await shoppingDeliveryAddressServices.GetShoppingDeliveryAddressAsync(id);
+            //check if true user
+            if (shoppingbyUser.UserId == this.UserId || Library.isAdmin(roleName))
             {
-                Success = true,
-                Message = new[] { "Delete Success" }
-            });
+                await shoppingDeliveryAddressServices.DeleteShoppingDeliveryAddress(id);
+                return Ok(new ResponseApi
+                {
+                    Success = true,
+                    Message = new[] { "Delete Success" }
+                });
+            }
+            return Unauthorized();
         }
     }
 }
